requirements.txt
alembic==1.4.2
attrs==19.3.0
blinker==1.4
boto3==1.14.0
botocore==1.17.0
certifi==2020.4.5.2
chardet==3.0.4
click==7.1.2
docutils==0.15.2
ecdsa==0.15
Flask==1.1.2
Flask-Cors==3.0.8
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.3
Flask-WhooshAlchemy==0.56
gunicorn==20.0.4
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
jsonschema==3.2.0
Mako==1.1.3
MarkupSafe==1.1.1
psycopg2-binary==2.8.5
pyasn1==0.4.8
pyrsistent==0.16.0
python-dateutil==2.8.1
python-dotenv==0.13.0
python-editor==1.0.4
python-jose==3.1.0
PyYAML==5.3.1
requests==2.23.0
requests-aws4auth==1.0
rsa==4.0
s3transfer==0.3.3
simplejson==3.17.0
six==1.15.0
SQLAlchemy==1.3.17
sqlathanor==0.5.1
urllib3==1.25.9
validator-collection==1.4.1
Werkzeug==1.0.1
Whoosh==2.7.4
psycopg2==2.7.3.1
